# storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SzCloudApi",
        "description": "API para almacenamiento en la nube — gestión de archivos, carpetas y papelera.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/expansions/{id}": {
            "get": {
                "tags": [
                    "Expansions"
                ],
                "summary": "Obtener info de una expansión",
                "description": "Retorna los detalles de una expansión específica.",
                "operationId": "317c17b0b3e85553d17dbb25914a561b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la expansión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la expansión"
                    },
                    "404": {
                        "description": "Expansión no encontrada"
                    }
                }
            }
        },
        "/api/expansions/{id}/buy": {
            "post": {
                "tags": [
                    "Expansions"
                ],
                "summary": "Comprar una expansión",
                "description": "Registra la compra de una expansión y suma el espacio al límite del usuario.",
                "operationId": "6ef3259054b3d2b151bdb393d844fce2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la expansión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expansión aplicada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Expansión no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación JWT. Usa el botón Authorize con email y contraseña para obtener el token automáticamente.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}